Move angular related parts from kibana_legacy to monitoring (#114977)
* Move angular related parts from kibana_legacy to monitoring

Closes: #114977

* remove private

* move format angular http error into monitoring

* fix translations
