+++ src/plugins/kibana_legacy/public/index.ts
@@ -16,7 +16,4 @@ export const plugin = (initializerContext: PluginInitializerContext) =>
   new KibanaLegacyPlugin(initializerContext);
 
 export * from './plugin';
-
-export * from './angular';
 export * from './notify';
-export * from './utils';
