+++ x-pack/plugins/monitoring/public/angular/helpers/format_angular_http_error.ts
@@ -1,15 +1,14 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
  */
 
 import { i18n } from '@kbn/i18n';
-import { IHttpResponse } from 'angular';
+import type { IHttpResponse } from 'angular';
 
-export type AngularHttpError = IHttpResponse<{ message: string }>;
+type AngularHttpError = IHttpResponse<{ message: string }>;
 
 export function isAngularHttpError(error: any): error is AngularHttpError {
   return (
@@ -25,7 +24,7 @@ export function formatAngularHttpError(error: AngularHttpError) {
   // is an Angular $http "error object"
   if (error.status === -1) {
     // status = -1 indicates that the request was failed to reach the server
-    return i18n.translate('kibana_legacy.notify.fatalError.unavailableServerErrorMessage', {
+    return i18n.translate('xpack.monitoring.notify.fatalError.unavailableServerErrorMessage', {
       defaultMessage:
         'An HTTP request has failed to connect. ' +
         'Please check if the Kibana server is running and that your browser has a working connection, ' +
@@ -33,7 +32,7 @@ export function formatAngularHttpError(error: AngularHttpError) {
     });
   }
 
-  return i18n.translate('kibana_legacy.notify.fatalError.errorStatusMessage', {
+  return i18n.translate('xpack.monitoring.notify.fatalError.errorStatusMessage', {
     defaultMessage: 'Error {errStatus} {errStatusText}: {errMessage}',
     values: {
       errStatus: error.status,
