+++ x-pack/plugins/monitoring/public/angular/top_nav/angular_config.tsx
@@ -1,9 +1,8 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
  */
 
 import {
@@ -23,7 +22,7 @@ import { ChromeBreadcrumb, EnvironmentMode, PackageInfo } from 'kibana/public';
 import { History } from 'history';
 
 import { CoreStart } from 'kibana/public';
-import { formatAngularHttpError, isAngularHttpError } from '../notify/lib';
+import { formatAngularHttpError, isAngularHttpError } from '../helpers/format_angular_http_error';
 
 export interface RouteConfiguration {
   controller?: string | ((...args: any[]) => void);
