+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -4221,8 +4221,6 @@
     "inspector.requests.statisticsTabLabel": "统计信息",
     "inspector.title": "检查器",
     "inspector.view": "视图：{viewName}",
-    "kibana_legacy.notify.fatalError.errorStatusMessage": "错误 {errStatus} {errStatusText}：{errMessage}",
-    "kibana_legacy.notify.fatalError.unavailableServerErrorMessage": "HTTP 请求无法连接。请检查 Kibana 服务器是否正在运行以及您的浏览器是否具有有效的连接，或请联系您的系统管理员。",
     "kibana_legacy.notify.toaster.errorStatusMessage": "错误 {errStatus} {errStatusText}：{errMessage}",
     "kibana_legacy.notify.toaster.unavailableServerErrorMessage": "HTTP 请求无法连接。请检查 Kibana 服务器是否正在运行以及您的浏览器是否具有有效的连接，或请联系您的系统管理员。",
     "kibana_utils.history.savedObjectIsMissingNotificationMessage": "已保存对象缺失",
